Make NessieClient an interface
