+++ clients/client/src/main/java/org/projectnessie/client/http/HttpRequest.java
@@ -27,6 +27,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.function.BiConsumer;
 
 import javax.net.ssl.HttpsURLConnection;
 import javax.net.ssl.SSLContext;
@@ -90,31 +91,45 @@ public class HttpRequest {
         ((HttpsURLConnection) con).setSSLSocketFactory(sslContext.getSocketFactory());
       }
       RequestContext context = new RequestContext(headers, uri, method, body);
-      requestFilters.forEach(a -> a.filter(context));
-      headers.entrySet()
-             .stream()
-             .flatMap(e -> e.getValue().stream().map(x -> new SimpleImmutableEntry<>(e.getKey(), x)))
-             .forEach(x -> con.setRequestProperty(x.getKey(), x.getValue()));
-      con.setRequestMethod(method.name());
-      if (method.equals(Method.PUT) || method.equals(Method.POST)) {
-        // Need to set the Content-Type even if body==null, otherwise the server responds with
-        // RESTEASY003065: Cannot consume content type
-        con.setRequestProperty("Content-Type", "application/json; charset=utf-8");
-        if (body != null) {
-          con.setDoOutput(true);
-          Class<?> bodyType = body.getClass();
-          if (bodyType != String.class) {
-            mapper.writerFor(bodyType).writeValue(con.getOutputStream(), body);
-          } else {
-            // This is mostly used for testing bad/broken JSON
-            con.getOutputStream().write(((String) body).getBytes(StandardCharsets.UTF_8));
+      ResponseContext responseContext = new ResponseContextImpl(con);
+      try {
+        requestFilters.forEach(a -> a.filter(context));
+        headers.entrySet()
+               .stream()
+               .flatMap(e -> e.getValue().stream().map(x -> new SimpleImmutableEntry<>(e.getKey(), x)))
+               .forEach(x -> con.setRequestProperty(x.getKey(), x.getValue()));
+        con.setRequestMethod(method.name());
+        if (method.equals(Method.PUT) || method.equals(Method.POST)) {
+          // Need to set the Content-Type even if body==null, otherwise the server responds with
+          // RESTEASY003065: Cannot consume content type
+          con.setRequestProperty("Content-Type", "application/json; charset=utf-8");
+          if (body != null) {
+            con.setDoOutput(true);
+            Class<?> bodyType = body.getClass();
+            if (bodyType != String.class) {
+              mapper.writerFor(bodyType).writeValue(con.getOutputStream(), body);
+            } else {
+              // This is mostly used for testing bad/broken JSON
+              con.getOutputStream().write(((String) body).getBytes(StandardCharsets.UTF_8));
+            }
           }
         }
+        con.connect();
+        con.getResponseCode(); // call to ensure http request is complete
+
+        List<BiConsumer<ResponseContext, Exception>> callbacks = context.getResponseCallbacks();
+        if (callbacks != null) {
+          callbacks.forEach(callback -> callback.accept(responseContext, null));
+        }
+      } catch (IOException e) {
+        List<BiConsumer<ResponseContext, Exception>> callbacks = context.getResponseCallbacks();
+        if (callbacks != null) {
+          callbacks.forEach(callback -> callback.accept(null, e));
+        }
       }
-      con.connect();
-      con.getResponseCode(); // call to ensure http request is complete
-      ResponseContext responseContext = new ResponseContextImpl(con);
-      responseFilters.forEach(a -> a.filter(responseContext));
+
+      responseFilters.forEach(responseFilter -> responseFilter.filter(responseContext));
+
       return new HttpResponse(responseContext, mapper);
     } catch (ProtocolException e) {
       throw new HttpClientException(String.format("Cannot perform request. Invalid protocol %s", method), e);
